refactor: overlay usage
