+++ src/components/ProductModal/ProductModal.module.scss
@@ -67,8 +67,4 @@
 		padding: var(--s-3);
 		width: 95%;
 	}
-
-	.dialog {
-		padding-inline: var(--s-1);
-	}
 }
